Move card reader to second core
Most of the time engaging with the card reader is not a problem on the
main core, but when the card is not placed properly on the reader it is
possible that it requires quite a few tries to obtain the card id which
takes too long and leads to audio glitches. Doing this on core 1 is the
safe way to always avoid such glitches.