Replace RefCell hack with ManuallyDrop hack
The problem why we need it at all is that borrowck denies us reborrowing
the the SDCardController because SDCardFile has a drop impl. The
previous "solution" worked around this by only taking a shared reference
(to a RefCell for interior mutability). As it turns out the problem is
the Drop impl on SDCardFile which we "remove" using a ManuallyDrop
wrapper and dropping the contents manually in the appropriate places.
Hopefully this gets fixed in a future compiler version.