Try to voltage-spike-proof on/off mechanism
Large fluctuations in the power (e.g. due to an electrical motor
starting/stopping) may cause the power to the microcontroller to drop
which apparently causes the gpio interrupt to fire spuriously. The
debouncing procedure is changed to check the pin state after the
debouncing period and the sleep is reentered if no actual press (after
debouncing) was registered.